requirements.txt
PyYAML>=3.12
allure-pytest>=2.6.0